app/bigquery/dataset/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
requests==2.31.0
google-cloud-bigquery==3.17.1
asyncio==3.4.3
